// features/api/model/nozzlesIndex200.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */
import type { NozzleResource } from './nozzleResource';
import type { NozzlesIndex200Links } from './nozzlesIndex200Links';
import type { NozzlesIndex200Meta } from './nozzlesIndex200Meta';

export type NozzlesIndex200 = {
  data: NozzleResource[];
  links: NozzlesIndex200Links;
  meta: NozzlesIndex200Meta;
};